+++ src/main/java/com/romanmarkunas/blog/memory/example15/OrderView.java
@@ -10,16 +10,16 @@ public final class OrderView {
 
     private final static int ID_OFFSET = 0;
     private final static int USER_OFFSET = ID_OFFSET + Long.BYTES;
-    private final static int ARTICLENR_OFFSET = USER_OFFSET + Long.BYTES;
+    private final static int ARTICLENR_OFFSET = USER_OFFSET + Integer.BYTES;
     private final static int COUNT_OFFSET = ARTICLENR_OFFSET + Integer.BYTES;
     private final static int PRICEPENCE_OFFSET = COUNT_OFFSET + Integer.BYTES;
     private final static int ADDRESSNUMBER_OFFSET = PRICEPENCE_OFFSET + Integer.BYTES;
-    private final static int ADDRESSSTREET_OFFSET = ADDRESSNUMBER_OFFSET + Long.BYTES;
-    private final static int ADDRESSCITY_OFFSET = ADDRESSSTREET_OFFSET + Long.BYTES;
-    private final static int ADDRESSREGION_OFFSET = ADDRESSCITY_OFFSET + Long.BYTES;
-    private final static int ADDRESSPOSTCODE_OFFSET = ADDRESSREGION_OFFSET + Long.BYTES;
+    private final static int ADDRESSSTREET_OFFSET = ADDRESSNUMBER_OFFSET + Integer.BYTES;
+    private final static int ADDRESSCITY_OFFSET = ADDRESSSTREET_OFFSET + Integer.BYTES;
+    private final static int ADDRESSREGION_OFFSET = ADDRESSCITY_OFFSET + Integer.BYTES;
+    private final static int ADDRESSPOSTCODE_OFFSET = ADDRESSREGION_OFFSET + Integer.BYTES;
 
-    final static int TOTAL_SIZE = ADDRESSPOSTCODE_OFFSET + Long.BYTES;
+    final static int TOTAL_SIZE = ADDRESSPOSTCODE_OFFSET + Integer.BYTES;
 
     private final PooledByteArrayMap byteArrayPool;
 
@@ -50,15 +50,15 @@ public final class OrderView {
             String addressPostCode
     ) {
         buffer.putLong(startPosition + ID_OFFSET, id);
-        buffer.putLong(startPosition + USER_OFFSET, byteArrayPool.put(user));
+        buffer.putInt(startPosition + USER_OFFSET, byteArrayPool.put(user));
         buffer.putInt(startPosition + ARTICLENR_OFFSET, articleNr);
         buffer.putInt(startPosition + COUNT_OFFSET, count);
         buffer.putInt(startPosition + PRICEPENCE_OFFSET, pricePence);
-        buffer.putLong(startPosition + ADDRESSNUMBER_OFFSET, internString(addressNumber));
-        buffer.putLong(startPosition + ADDRESSSTREET_OFFSET, internString(addressStreet));
-        buffer.putLong(startPosition + ADDRESSCITY_OFFSET, internString(addressCity));
-        buffer.putLong(startPosition + ADDRESSREGION_OFFSET, internString(addressRegion));
-        buffer.putLong(startPosition + ADDRESSPOSTCODE_OFFSET, internString(addressPostCode));
+        buffer.putInt(startPosition + ADDRESSNUMBER_OFFSET, internString(addressNumber));
+        buffer.putInt(startPosition + ADDRESSSTREET_OFFSET, internString(addressStreet));
+        buffer.putInt(startPosition + ADDRESSCITY_OFFSET, internString(addressCity));
+        buffer.putInt(startPosition + ADDRESSREGION_OFFSET, internString(addressRegion));
+        buffer.putInt(startPosition + ADDRESSPOSTCODE_OFFSET, internString(addressPostCode));
         return this;
     }
 
@@ -71,8 +71,8 @@ public final class OrderView {
         return byteArrayPool.get(getUserPoolKey());
     }
 
-    public long getUserPoolKey() {
-        return buffer.getLong(startPosition + USER_OFFSET);
+    public int getUserPoolKey() {
+        return buffer.getInt(startPosition + USER_OFFSET);
     }
 
     public int getArticleNr() {
@@ -88,62 +88,62 @@ public final class OrderView {
     }
 
     public String getAddressNumber() {
-        long addressNumber = buffer.getLong(startPosition + ADDRESSNUMBER_OFFSET);
+        int addressNumber = buffer.getInt(startPosition + ADDRESSNUMBER_OFFSET);
         return restoreString(addressNumber);
     }
 
     public OrderView changeAddressNumber(String addressNumber) {
-        long oldAddressNumber = buffer.getLong(startPosition + ADDRESSNUMBER_OFFSET);
-        long newAddressNumber = replaceString(addressNumber, oldAddressNumber);
-        buffer.putLong(startPosition + ADDRESSNUMBER_OFFSET, newAddressNumber);
+        int oldAddressNumber = buffer.getInt(startPosition + ADDRESSNUMBER_OFFSET);
+        int newAddressNumber = replaceString(addressNumber, oldAddressNumber);
+        buffer.putInt(startPosition + ADDRESSNUMBER_OFFSET, newAddressNumber);
         return this;
     }
 
     public String getAddressStreet() {
-        long addressStreet = buffer.getLong(startPosition + ADDRESSSTREET_OFFSET);
+        int addressStreet = buffer.getInt(startPosition + ADDRESSSTREET_OFFSET);
         return restoreString(addressStreet);
     }
 
     public OrderView changeAddressStreet(String addressStreet) {
-        long oldAddressStreet = buffer.getLong(startPosition + ADDRESSSTREET_OFFSET);
-        long newAddressStreet = replaceString(addressStreet, oldAddressStreet);
-        buffer.putLong(startPosition + ADDRESSSTREET_OFFSET, newAddressStreet);
+        int oldAddressStreet = buffer.getInt(startPosition + ADDRESSSTREET_OFFSET);
+        int newAddressStreet = replaceString(addressStreet, oldAddressStreet);
+        buffer.putInt(startPosition + ADDRESSSTREET_OFFSET, newAddressStreet);
         return this;
     }
 
     public String getAddressCity() {
-        long addressCity = buffer.getLong(startPosition + ADDRESSCITY_OFFSET);
+        int addressCity = buffer.getInt(startPosition + ADDRESSCITY_OFFSET);
         return restoreString(addressCity);
     }
 
     public OrderView changeAddressCity(String addressCity) {
-        long oldAddressCity = buffer.getLong(startPosition + ADDRESSCITY_OFFSET);
-        long newAddressCity = replaceString(addressCity, oldAddressCity);
-        buffer.putLong(startPosition + ADDRESSCITY_OFFSET, newAddressCity);
+        int oldAddressCity = buffer.getInt(startPosition + ADDRESSCITY_OFFSET);
+        int newAddressCity = replaceString(addressCity, oldAddressCity);
+        buffer.putInt(startPosition + ADDRESSCITY_OFFSET, newAddressCity);
         return this;
     }
 
     public String getAddressRegion() {
-        long addressRegion = buffer.getLong(startPosition + ADDRESSREGION_OFFSET);
+        int addressRegion = buffer.getInt(startPosition + ADDRESSREGION_OFFSET);
         return restoreString(addressRegion);
     }
 
     public OrderView changeAddressRegion(String addressRegion) {
-        long oldAddressRegion = buffer.getLong(startPosition + ADDRESSREGION_OFFSET);
-        long newAddressRegion = replaceString(addressRegion, oldAddressRegion);
-        buffer.putLong(startPosition + ADDRESSREGION_OFFSET, newAddressRegion);
+        int oldAddressRegion = buffer.getInt(startPosition + ADDRESSREGION_OFFSET);
+        int newAddressRegion = replaceString(addressRegion, oldAddressRegion);
+        buffer.putInt(startPosition + ADDRESSREGION_OFFSET, newAddressRegion);
         return this;
     }
 
     public String getAddressPostCode() {
-        long addressPostCode = buffer.getLong(startPosition + ADDRESSPOSTCODE_OFFSET);
+        int addressPostCode = buffer.getInt(startPosition + ADDRESSPOSTCODE_OFFSET);
         return restoreString(addressPostCode);
     }
 
     public OrderView changeAddressPostCode(String addressPostCode) {
-        long oldAddressPostCode = buffer.getLong(startPosition + ADDRESSPOSTCODE_OFFSET);
-        long newAddressPostCode = replaceString(addressPostCode, oldAddressPostCode);
-        buffer.putLong(startPosition + ADDRESSPOSTCODE_OFFSET, newAddressPostCode);
+        int oldAddressPostCode = buffer.getInt(startPosition + ADDRESSPOSTCODE_OFFSET);
+        int newAddressPostCode = replaceString(addressPostCode, oldAddressPostCode);
+        buffer.putInt(startPosition + ADDRESSPOSTCODE_OFFSET, newAddressPostCode);
         return this;
     }
 
@@ -156,19 +156,18 @@ public final class OrderView {
     }
 
 
-    private long internString(String str) {
+    private int internString(String str) {
         return byteArrayPool.put(str.getBytes(StandardCharsets.US_ASCII));
     }
 
-    private long replaceString(String str, long oldKey) {
-        long newKey = internString(str);
+    private int replaceString(String str, int oldKey) {
+        int newKey = internString(str);
         byteArrayPool.free(oldKey);
         return newKey;
     }
 
     @SuppressWarnings("byte.array.weakening")
-    private String restoreString(long key) {
+    private String restoreString(int key) {
         return new String(byteArrayPool.get(key), StandardCharsets.US_ASCII);
     }
-
 }
